Move `Assets<T>` to resources
