+++ src/asset/core/index.js
@@ -1,4 +1,3 @@
-export * from './asset.js'
 export * from './channel.js'
 export * from './handle.js'
 export * from './exporter.js'
