+++ src/asset/plugins/asset.js
@@ -4,7 +4,7 @@ import { App, Plugin } from '../../app/index.js'
 import { AppSchedule, CoreSystems } from '../../core/index.js'
 import { EventPlugin } from '../../event/index.js'
 import { typeid, typeidGeneric } from '../../type/index.js'
-import { Assets } from '../core/index.js'
+import { Assets } from '../resources/index.js'
 import { AssetAdded, AssetDropped, AssetModified } from '../events/index.js'
 import { registerAssetTypes, registerAssetOnAssetServer, unloadDroppedAssets, updateAssetChannel, updateAssetEvents } from '../systems/index.js'
