+++ src/asset/resources/index.js
@@ -1,1 +1,2 @@
+export * from './assets.js'
 export * from './assetserver.js'
